infra/lambda-functions/file-processor/requirements.txt
langchain-core
langchain_community
langchain-openai
pytz
pydantic